Implement the CloudService GRPC, which is used by Minion, and remove RpcRequest GRPC which is not used by Minion.
